update: 创建任务选择节点页面重构
