+++ frontend/desktop/src/components/common/TemplateCanvas/NodeTemplate/TaskNode.vue
@@ -27,6 +27,7 @@
                     <bk-checkbox
                         v-else-if="node.mode === 'select'"
                         :value="node.checked"
+                        :disabled="node.checkDisable"
                         @change="onNodeCheckClick">
                     </bk-checkbox>
                 </template>
@@ -139,6 +140,9 @@
                 this.$emit('onModifyTimeClick', this.node.id)
             },
             onNodeCheckClick () {
+                if (this.node.checkDisable) {
+                    return
+                }
                 this.$emit('onNodeCheckClick', this.node.id, !this.node.checked)
             }
         }
