+++ frontend/desktop/src/pages/task/TaskCreate/TaskSelectNode.vue
@@ -12,74 +12,6 @@
 <template>
     <div class="select-node-wrapper" v-bkloading="{ isLoading: loading, opacity: 1 }">
         <div class="canvas-content">
-            <div
-                v-if="isSchemeShow"
-                :class="[
-                    'scheme-right-header',
-                    { 'scheme-toggle-right-header': !showPanel }
-                ]">
-                <div class="scheme-combine-shape" @click="toggleSchemePanel">
-                    <i class="common-icon-paper" v-bk-tooltips="{
-                        content: i18n.schema,
-                        placements: ['top']
-                    }"></i>
-                </div>
-            </div>
-            <div class="node-select-scheme" v-if="isSchemeShow && showPanel">
-                <div class="scheme-title">
-                    <span> {{i18n.schemaList}}</span>
-                </div>
-                <div class="scheme-header">
-                    <div class="scheme-form" v-if="taskActionShow">
-                        <bk-input
-                            v-model="schemeName"
-                            v-validate="schemeNameRule"
-                            data-vv-validate-on=" "
-                            name="schemeName"
-                            class="bk-input-inline"
-                            :clearable="true"
-                            :placeholder="i18n.schemaName">
-                        </bk-input>
-                        <bk-button theme="success" @click="onAddScheme">{{i18n.affirm}}</bk-button>
-                        <bk-button @click="onCancelScheme">{{i18n.actionCancel}}</bk-button>
-                        <span v-if="errors.has('schemeName')" class="common-error-tip error-msg">{{ errors.first('schemeName') }}</span>
-                    </div>
-                    <bk-button
-                        v-else
-                        theme="primary"
-                        :class="['save-scheme-btn', {
-                            'disabled-btn': isPreviewMode,
-                            'btn-permission-disable': !hasPermission(['create_scheme'], tplActions, tplOperations)
-                        }]"
-                        @click="onShowSchemeDialog">
-                        {{ i18n.newSchema }}
-                    </bk-button>
-                </div>
-                <div class="scheme-content">
-                    <ul class="schemeList">
-                        <li
-                            v-for="item in taskScheme"
-                            :class="{
-                                'scheme-item': true,
-                                'selected': item.id === selectedScheme && lastSelectSchema === item.id
-                            }"
-                            :key="item.id"
-                            @click="onSelectScheme(item.id)">
-                            <a class="scheme-name" :title="item.name">{{item.name}}</a>
-                            <i class="bk-icon icon-close-circle-shape" @click.stop="onDeleteScheme(item.id)"></i>
-                        </li>
-                    </ul>
-                </div>
-                <div class="scheme-preview-mode">
-                    <div class="scheme-header-division-line scheme-header-division-line-last"></div>
-                    <div class="preview-mode-switcher">
-                        <span>
-                            {{i18n.previewMode}}
-                        </span>
-                        <bk-switcher size="small" v-model="isPreviewMode" @change="onChangePreviewNode"></bk-switcher>
-                    </div>
-                </div>
-            </div>
             <TemplateCanvas
                 v-if="!loading && !isPreviewMode"
                 ref="templateCanvas"
@@ -99,12 +31,22 @@
                 :preview-data-loading="previewDataLoading"
                 :canvas-data="formatCanvasData('perview', previewData)"
                 :preview-bread="previewBread"
-                :is-all-selected="isAllSelected"
-                :is-show-select-all-tool="viewMode !== 'appmaker'"
-                :is-select-all-tool-disabled="isPreviewMode"
                 @onNodeClick="onNodeClick"
                 @onSelectSubflow="onSelectSubflow">
             </NodePreview>
+            <task-scheme
+                :project_id="project_id"
+                :template_id="template_id"
+                :template-name="templateName"
+                :is-scheme-show="isSchemeShow"
+                :is-preview-mode="isPreviewMode"
+                :selected-nodes="selectedNodes"
+                :tpl-actions="tplActions"
+                :tpl-operations="tplOperations"
+                :tpl-resource="tplResource"
+                @selectScheme="selectScheme"
+                @togglePreviewMode="togglePreviewMode">
+            </task-scheme>
         </div>
         <div class="action-wrapper" slot="action-wrapper">
             <bk-button
@@ -120,48 +62,29 @@
     import { mapState, mapMutations, mapActions } from 'vuex'
     import { errorHandler } from '@/utils/errorHandler.js'
     import tools from '@/utils/tools.js'
-    import { NAME_REG, STRING_LENGTH } from '@/constants/index.js'
+    import TaskScheme from './TaskScheme.vue'
     import TemplateCanvas from '@/components/common/TemplateCanvas/index.vue'
     import NodePreview from '@/pages/task/NodePreview.vue'
-    import permission from '@/mixins/permission.js'
 
     export default {
         name: 'TaskSelectNode',
         components: {
+            TaskScheme,
             TemplateCanvas,
             NodePreview
         },
-        mixins: [permission],
         props: ['project_id', 'template_id', 'common', 'excludeNode', 'entrance'],
         data () {
             return {
                 i18n: {
-                    schemaList: gettext('执行方案列表'),
-                    newSchema: gettext('新建'),
                     all: gettext('全选'),
-                    affirm: gettext('保存'),
                     cancel: gettext('取消全选'),
-                    actionCancel: gettext('取消'),
-                    next: gettext('下一步'),
-                    save: gettext('执行方案保存'),
-                    schemaName: gettext('方案名称'),
-                    previewMode: gettext('预览模式：'),
-                    schema: gettext('执行方案')
+                    next: gettext('下一步')
                 },
                 loading: true,
-                bkMessageInstance: null,
-                isSubmit: false,
-                isDelete: false,
-                showPanel: true,
                 selectedNodes: [],
                 allSelectableNodes: [],
                 selectedScheme: '',
-                schemeName: '',
-                schemeNameRule: {
-                    required: true,
-                    max: STRING_LENGTH.SCHEME_NAME_MAX_LENGTH,
-                    regex: NAME_REG
-                },
                 isPreviewMode: false,
                 previewDataLoading: false,
                 version: '',
@@ -172,15 +95,13 @@
                     gateways: {},
                     constants: []
                 },
-                taskName: '',
+                templateName: '',
                 pipelineData: '',
                 isPreview: false,
-                lastSelectSchema: '',
                 isAllSelected: true,
-                taskActionShow: false,
                 tplActions: [],
                 tplOperations: [],
-                tplResource: []
+                tplResource: {}
             }
         },
         computed: {
@@ -218,9 +139,6 @@
                 'getLayoutedPipeline'
             ]),
             ...mapActions('task/', [
-                'loadTaskScheme',
-                'createTaskScheme',
-                'deleteTaskScheme',
                 'getSchemeDetail',
                 'loadPreviewNodeData'
             ]),
@@ -229,9 +147,6 @@
             ]),
             ...mapMutations('template/', [
                 'setTemplateData'
-            ]),
-            ...mapMutations('task/', [
-                'setTaskScheme'
             ]),
             /**
              * 获取模板数据，并设置至store中
@@ -249,15 +164,13 @@
                     this.tplOperations = templateData.auth_operations
                     this.tplResource = templateData.auth_resource
                     this.version = templateData.version
-                    this.taskName = templateData.name
-                    const schemeData = await this.loadTaskScheme({ 'project_id': this.project_id, 'template_id': this.template_id, 'isCommon': this.isCommonProcess })
+                    this.templateName = templateData.name
+
                     if (this.viewMode === 'appmaker') {
                         const appmakerData = await this.loadAppmakerDetail(this.app_id)
                         const schemeId = Number(appmakerData.template_scheme_id)
                         schemeId && this.onSelectScheme(schemeId)
                     }
-                
-                    this.setTaskScheme(schemeData)
                     this.setTemplateData(templateData)
                     this.allSelectableNodes = this.location.filter(item => item.optional)
                     this.allSelectableNodes.forEach(item => {
@@ -284,7 +197,6 @@
              */
             async getPreviewNodeData (templateId, isSubflow = false, inExcludeNode) {
                 this.previewDataLoading = true
-                this.isPreview = true
                 const excludeNode = isSubflow ? [] : this.getExcludeNode()
                 const templateSource = this.common ? 'common' : 'business'
                 const params = {
@@ -335,159 +247,17 @@
                 }
             },
             /**
-             * 任务方案面板是否显示
-             */
-            toggleSchemePanel () {
-                this.showPanel = !this.showPanel
-            },
-            /**
-             * 创建任务方案弹窗
-             */
-            onShowSchemeDialog () {
-                if (!this.hasPermission(['create_scheme'], this.tplActions, this.tplOperations)) {
-                    const resourceData = {
-                        name: this.taskName,
-                        id: this.template_id,
-                        auth_actions: this.tplActions
-                    }
-                    this.applyForPermission(['create_scheme'], resourceData, this.tplOperations, this.tplResource)
-                    return
-                }
-                if (!this.isPreviewMode) {
-                    this.taskActionShow = true
-                }
-            },
-            getExcludeNode () {
-                const nodes = []
-                this.allSelectableNodes.filter(item => {
-                    if (this.selectedNodes.indexOf(item.id) === -1) {
-                        nodes.push(item.id)
-                    }
-                })
-                return nodes
-            },
-            /**
-             * 添加方案
-             */
-            onAddScheme () {
-                if (this.isSubmit) return
-                const isSchemeNameExist = this.taskScheme.some(item => item.name === this.schemeName)
-                if (isSchemeNameExist) {
-                    errorHandler({ message: gettext('方案名称已存在') }, this)
-                    return
-                }
-                this.isSubmit = true
-                this.$validator.validateAll().then(async (result) => {
-                    if (!result) {
-                        this.isSubmit = false
-                        return
-                    }
-                    this.schemeName = this.schemeName.trim()
-                    const selectedNodes = this.selectedNodes.slice()
-                    const scheme = {
-                        project_id: this.project_id,
-                        template_id: this.template_id,
-                        name: this.schemeName,
-                        data: JSON.stringify(selectedNodes),
-                        isCommon: this.isCommonProcess
-                    }
-                    try {
-                        const newScheme = await this.createTaskScheme(scheme)
-                        const schemeData = await this.loadTaskScheme({
-                            'project_id': this.project_id,
-                            'template_id': this.template_id,
-                            'isCommon': this.isCommonProcess
-                        })
-                        this.setTaskScheme(schemeData)
-                        this.selectedScheme = newScheme.id
-                        this.lastSelectSchema = newScheme.id
-                        this.schemeName = ''
-                        this.taskActionShow = false
-                        this.$bkMessage({
-                            message: gettext('方案添加成功'),
-                            theme: 'success'
-                        })
-                    } catch (e) {
-                        errorHandler(e, this)
-                    } finally {
-                        this.isSubmit = false
-                    }
-                })
-            },
-            /**
-             * 选择方案并进行切换更新选择的节点
-             */
-            async onSelectScheme (id) {
-                this.selectedScheme = id
-                if (this.lastSelectSchema === id) {
-                    this.lastSelectSchema = ''
-                    if (this.isPreviewMode) {
-                        await this.getPreviewNodeData(this.template_id, true)
-                    }
-                } else {
-                    this.lastSelectSchema = id
-                    try {
-                        const data = await this.getSchemeDetail({ id: id, isCommon: this.isCommonProcess })
-                        this.selectedNodes = tools.deepClone(data.data)
-                        const excludeNode = this.getExcludeNode()
-                        this.$emit('setExcludeNode', excludeNode)
-                        this.canvasData.locations.forEach(item => {
-                            if (this.isSelectableNode(item.id)) {
-                                const checked = this.selectedNodes.indexOf(item.id) > -1
-                                this.$set(item, 'checked', checked)
-                            }
-                        })
-                        if (this.isPreviewMode) {
-                            await this.getPreviewNodeData(this.template_id, false, excludeNode)
-                        }
-                    } catch (e) {
-                        errorHandler(e, this)
-                    }
-                }
-            },
-            /**
-             * 删除方案
-             */
-            async onDeleteScheme (id) {
-                if (!this.hasPermission(['delete_scheme'], this.tplActions, this.tplOperations)) {
-                    const resourceData = {
-                        name: this.taskName,
-                        id: this.template_id,
-                        auth_actions: this.tplActions
-                    }
-                    this.applyForPermission(['delete_scheme'], resourceData, this.tplOperations, this.tplResource)
-                    return
-                }
-
-                if (this.isDelete) return
-                this.isDelete = true
-                try {
-                    await this.deleteTaskScheme({ id: id, isCommon: this.isCommonProcess })
-                    const schemeData = await this.loadTaskScheme({ 'project_id': this.project_id, 'template_id': this.template_id, isCommon: this.isCommonProcess })
-                    this.setTaskScheme(schemeData)
-                    this.$bkMessage({
-                        message: gettext('方案删除成功'),
-                        theme: 'success'
-                    })
-                } catch (e) {
-                    errorHandler(e, this)
-                } finally {
-                    this.isDelete = false
-                }
-            },
-            /**
              * 进入参数填写阶段，设置执行节点
              */
             async onGotoParamFill () {
                 this.loading = true
                 const excludeNode = this.getExcludeNode()
                 try {
-                    if (!this.isPreview) {
+                    if (!this.isPreviewMode) {
                         await this.getPreviewNodeData(this.template_id)
                     }
                 
                     this.$emit('setExcludeNode', excludeNode)
-                    this.$emit('setPreviewData', this.pipelineData)
                 
                     this.loading = false
                     if (this.viewMode === 'appmaker') {
@@ -509,122 +279,6 @@
                 } catch (e) {
                     errorHandler(e, this)
                 }
-            },
-            /**
-             * 预览模式的点击事件
-             * @params {Boolean} isPreview  是否是预览模式
-             */
-            onChangePreviewNode (isPreview) {
-                if (isPreview) {
-                    const excludeNode = this.getExcludeNode()
-                    this.$emit('setExcludeNode', excludeNode)
-                    this.isPreviewMode = true
-                    this.previewBread.push({
-                        data: this.template_id,
-                        name: this.taskName
-                    })
-                    this.getPreviewNodeData(this.template_id)
-                } else {
-                    this.isPreviewMode = false
-                    this.previewBread = []
-                }
-            },
-            /**
-             * 清空没有被选中的节点
-             * @params {Object} previewNodeData  整个pipelineTree
-             * @return {Object} previewNodeData  清空了需要使用的节点的pipelineTree
-             */
-            clearUnnecessaryGateway (previewNodeData) {
-                const { line, location } = previewNodeData
-                // 数据分组 用于获取线段的后置连线
-                const group = []
-                for (const nowLine of line) {
-                    const sourceId = nowLine.source.id
-                    const targetId = nowLine.target.id
-                    if (group[sourceId]) {
-                        group[sourceId].push(targetId)
-                    } else {
-                        group[sourceId] = [targetId]
-                    }
-                }
-                location.forEach((nowLocation) => {
-                    const { type, id } = nowLocation
-                    if (type === 'parallelgateway' || type === 'branchgateway') {
-                        // 用于判断是否是有需要删除的线和内容
-                        const lineArray = []
-                        // 检验Array的每一个项相同
-                        group[id].forEach((targetId) => {
-                            if (lineArray.indexOf(targetId) === -1) {
-                                lineArray.push(targetId)
-                            }
-                        })
-                        const lineArrayLength = lineArray.length
-                        const deleteLocation = location.find(location => location.id === lineArray[0])
-                        let deleteLocationType
-                        if (deleteLocation !== undefined) {
-                            deleteLocationType = deleteLocation.type
-                        }
-                        // 长度为1 需要删除节点和线段
-                        if (lineArrayLength === 1 && deleteLocationType !== 'subflow' && deleteLocationType !== 'tasknode') {
-                            // 汇聚网关的id
-                            const convergegatewayId = group[id][0]
-                            const convergegateway = location.find(location => location.id === convergegatewayId)
-                            if (!convergegateway || convergegateway.type !== 'convergegateway') {
-                                // 如果不是汇聚网关就不需要删除了
-                                return
-                            }
-                            // 获取后置节点
-                            const lastId = line.find((newLine) => newLine.source.id === convergegatewayId).target.id
-                            // 将并行或分支网关的前一个节点连线线段和汇聚网关的下一个节点连接
-                            const preId = line.find((newLine) => newLine.target.id === id).source.id
-                            line.find((newLine) => newLine.source.id === preId && newLine.target.id === id).target.id = lastId
-                            // 删除节点
-                            while (true) {
-                                // 并行网关的前节点
-                                const frontLineSourceIndex = line.findIndex(item => item.source.id === id)
-                                if (frontLineSourceIndex !== -1) {
-                                    line.splice(frontLineSourceIndex, 1)
-                                }
-                                // 汇聚网关的后节点
-                                const lastLineTargetIndex = line.findIndex(item => item.target.id === convergegatewayId)
-                                if (lastLineTargetIndex !== -1) {
-                                    line.splice(lastLineTargetIndex, 1)
-                                }
-                                // 汇聚网关的前节点
-                                const lastLineSourceIndex = line.findIndex(item => item.source.id === convergegatewayId)
-                                if (lastLineSourceIndex !== -1) {
-                                    line.splice(lastLineSourceIndex, 1)
-                                }
-                            
-                                if (frontLineSourceIndex === -1 && lastLineTargetIndex === -1 && lastLineSourceIndex === -1) {
-                                    break
-                                }
-                            }
-                            // 删除节点的内容
-                            const lastLocationIndex = location.findIndex(item => item.id === convergegatewayId)
-                            location.splice(lastLocationIndex, 1)
-                            const frontLocationIndex = location.findIndex(item => item.id === id)
-                            location.splice(frontLocationIndex, 1)
-                        // 获取节点
-                        } else {
-                            // 可能需要删除空的连线
-                            lineArray.forEach((newLocationId) => {
-                                const deleteLocation = location.find((newLocation) => newLocation.id === newLocationId && newLocation.type === 'convergegateway')
-                                if (deleteLocation) {
-                                    for (const locationId of group[id]) {
-                                        // 使用
-                                        if (locationId === newLocationId) {
-                                            const frontLineSourceIndex = line.findIndex(item => item.target.id === newLocationId && item.source.id === id)
-                                            line.splice(frontLineSourceIndex, 1)
-                                        }
-                                    }
-                                }
-                            })
-                        }
-                    }
-                })
-                // 数据内容更新
-                return Object.assign(previewNodeData, { line, location })
             },
             /**
              * 格式化pipelineTree的数据，只输出一部分数据
@@ -697,6 +351,9 @@
                 })
                 this.getPreviewNodeData(templateId, true)
             },
+            /**
+             * 选中节点
+             */
             onNodeCheckClick (id, val) {
                 this.canvasData.locations.some(item => {
                     if (item.id === id) {
@@ -728,20 +385,66 @@
                 }
                 this.previewBread.splice(index + 1, this.previewBread.length)
             },
-            // 取消添加执行方案
-            onCancelScheme () {
-                this.schemeName = ''
-                this.taskActionShow = false
+            getExcludeNode () {
+                const nodes = []
+                this.allSelectableNodes.filter(item => {
+                    if (this.selectedNodes.indexOf(item.id) === -1) {
+                        nodes.push(item.id)
+                    }
+                })
+                return nodes
             },
             isSelectableNode (id) {
                 return this.allSelectableNodes.findIndex(item => item.id === id) > -1
+            },
+            /**
+             * 选择执行方案
+             */
+            async selectScheme (scheme) {
+                if (scheme === undefined) {
+                    if (this.isPreviewMode) {
+                        await this.getPreviewNodeData(this.template_id, true)
+                    }
+                } else {
+                    try {
+                        const data = await this.getSchemeDetail({ id: scheme, isCommon: this.isCommonProcess })
+                        this.selectedNodes = JSON.parse(data.data)
+                        const excludeNode = this.getExcludeNode()
+                        this.$emit('setExcludeNode', excludeNode)
+                        this.canvasData.locations.forEach(item => {
+                            if (this.isSelectableNode(item.id)) {
+                                const checked = this.selectedNodes.indexOf(item.id) > -1
+                                this.$set(item, 'checked', checked)
+                            }
+                        })
+                        if (this.isPreviewMode) {
+                            await this.getPreviewNodeData(this.template_id, false, excludeNode)
+                        }
+                    } catch (e) {
+                        errorHandler(e, this)
+                    }
+                }
+            },
+            togglePreviewMode (isPreview) {
+                this.isPreviewMode = isPreview
+                if (isPreview) {
+                    const excludeNode = this.getExcludeNode()
+                    this.$emit('setExcludeNode', excludeNode)
+                    this.previewBread.push({
+                        data: this.template_id,
+                        name: this.templateName
+                    })
+                    this.getPreviewNodeData(this.template_id)
+                } else {
+                    this.previewBread = []
+                }
             }
         }
     }
 </script>
 <style lang="scss" scoped>
-@import '@/scss/mixins/scrollbar.scss';
 @import '@/scss/config.scss';
+
 .select-node-wrapper {
     height: calc(100% - 90px);
 }
@@ -756,216 +459,6 @@
     }
     .node-preview-wrapper {
         height: 100%;
-    }
-}
-.node-select-scheme {
-    position: absolute;
-    top: 0;
-    right: 56px;
-    width: 420px;
-    height: 100%;
-    background: $whiteDefault;
-    border-left: 1px solid $commonBorderColor;
-    box-shadow: 0 0 8px rgba(0, 0, 0, 0.15);
-    z-index: 4;
-    transition: right 0.5s ease-in-out;
-    .scheme-title {
-        height: 35px;
-        margin: 20px;
-        border-bottom: 1px solid #cacecb;
-    }
-    .scheme-header {
-        margin: 20px;
-        .scheme-form {
-            display: inline-block;
-            input {
-                width: 200px;
-            }
-        }
-        .save-scheme-btn {
-            width: 90px;
-            height: 32px;
-            line-height: 32px;
-            background-color: #ffffff;
-            border: 1px solid #c4c6cc;
-            border-radius: 2px;
-            color: #313238;
-            font-size: 14px;
-            font-weight: 400;
-        }
-        .disabled-btn {
-            opacity: 0.3;
-            cursor: not-allowed;
-        }
-        .base-input {
-            height: 32px;
-            line-height: 32px;
-            padding-bottom: 2px;
-        }
-    }
-    .scheme-content {
-        height: calc(100% - 127px- 63px);
-        overflow: hidden;
-        overflow-y: auto;
-        @include scrollbar;
-        .scheme-item {
-            margin: 0 20px;
-            height: 42px;
-            font-weight: 400;
-            line-height: 42px;
-            font-size: 14px;
-            cursor: pointer;
-            border-bottom: 1px solid #ebebeb;
-            &:hover {
-                margin: 0;
-                padding: 0 20px;
-                background-color: #d9e8f8;
-                .icon-close-circle-shape {
-                    opacity: 1;
-                }
-            }
-            &.selected {
-                margin: -1px 0 0 0;
-                padding: 0 20px;
-                background-color: #3a84ff;
-                .scheme-name {
-                    color: #ffffff;
-                }
-                .scheme-division-line {
-                    background-color: #3a84ff;
-                }
-                .icon-close-circle-shape {
-                    color: #ffffff;
-                    opacity: 1;
-                }
-            }
-            .scheme-name {
-                display: inline-block;
-                width: 240px;
-                overflow: hidden;
-                white-space: nowrap;
-                text-overflow: ellipsis;
-                color: #313238;
-            }
-            .icon-close-circle-shape {
-                float: right;
-                margin-top: 15px;
-                margin-right: 5px;
-                width: 12px;
-                height: 12px;
-                text-align: center;
-                line-height: 12px;
-                color: #979ba5;
-                opacity: 0;
-                cursor: pointer;
-                &:hover {
-                    color: #cecece;
-                }
-            }
-        }
-        li:first-child {
-            border-top: 1px solid $commonBorderColor;
-        }
-    }
-    .scheme-preview-mode {
-        position: relative;
-        width: 420px;
-        .scheme-header-division-line-last {
-            margin: 0 25px 0 20px;
-            border: 0;
-            height: 1px;
-            background-color:#cacedb;
-        }
-        .preview-mode-switcher {
-            position: relative;
-            top: 19px;
-            left: 20px;
-            span {
-                font-size: 14px;
-                font-weight: 400;
-                color: #313238;
-            }
-        }
-    }
-}
-.quick-select {
-    position: absolute;
-    top: 15px;
-    left: 10px;
-    z-index: 5;
-}
-.toggle-button {
-    position: absolute;
-    top: 0;
-    left: -20px;
-    width: 20px;
-    height: 50px;
-    line-height: 50px;
-    color: $whiteDefault;
-    background: $blueThinBg;
-    text-align: center;
-    cursor: pointer;
-    &:hover {
-        background: $blueDefault;
-    }
-    .common-icon-arrow-left {
-        display: inline-block;
-        &.arrow-right {
-            transform: rotate(180deg);
-        }
-    }
-}
-
-.scheme-right-header {
-    position: absolute;
-    right: 0;
-    float: right;
-    width: 56px;
-    background: #ffffff;
-    border-left: 1px solid #cacedb;
-    height: 100%;
-    z-index: 5;
-    .scheme-combine-shape {
-        margin: 27px 0 0 12px;
-        width: 32px;
-        height: 32px;
-        background-color: #525F77;
-        border-radius:2px;
-        text-align: center;
-        line-height: 32px;
-        cursor: pointer;
-        .common-icon-paper {
-            color: #ffffff;
-        }
-    }
-}
-.scheme-toggle-right-header {
-    height: 86px;
-    border: 1px solid #cacedb;
-    border-top: 0px;
-    .scheme-combine-shape {
-        background-color: #ffffff;
-        .common-icon-paper {
-            color: #546a9e;
-        }
-    }
-}
-.disable-item {
-    cursor: not-allowed;
-    &:hover {
-        background: inherit ;
-    }
-}
-.scheme-name-wrapper {
-    padding: 10px 0;
-    label {
-        float: left;
-        margin-top: 6px;
-        width: 100px;
-        text-align: right;
-    }
-    .scheme-name-input {
-        margin: 0 35px 0 120px;
     }
 }
 .next-button {
@@ -991,9 +484,5 @@
         top: 19px;
         left: 40px;
     }
-}
-.bk-input-inline {
-    display: inline-block;
-    width: 200px;
 }
 </style>
